mortalityHazard

representation of a mortality hazard function

name

the hazard's name

num.param

the number of parameters the hazard uses

theta.default

default values for the parameters

theta.start.fn

if not NULL, a function which takes a mortalityData object and returns starting values to be used in a numerical optimization routine

binomial.grad.fn

if not NULL, a function which takes a parametr vector and a mortalityData object and returns the gradient of the binomial log-likelihood. Used in some numerical optimization routines

theta.range

if not NULL, a list with one entry for each parameter. Each parameter's entry has a range of plausible values for the parameter, which can be used if an optimization routine wants to choose a random starting value

optim.default

default settings to use with R's optim function

haz.fn

a function which takes two arguments, a vector of parameter values and a vector of ages, and returns the value of the hazard function with the given parameter values at each age.

haz.to.prob.fn

if not NULL a function which takes two arguments, a vector of parameter values and a vector of ages, and returns the value of \(exp(-\int_{z}^{z+1} -\mu(x)dx)\), ie the result of converting the hazard to conditional probabilities of dying between ages z and z+1 for each age z.

##'