mortalityModel

this object represents a likelihood function for use with mortality data. generally, it will be the result of combining a likelihood with a hazard; for example, it could be the binomial model with the gompertz hazard embedded in it. many of the slots of this object correspond to those found in mortalityHazard-class

Details

name

the model's name

loglik.fn

the log-likelihood function, which takes as its arguments

theta

a vector of parameter values

Dx

vector of deaths by age

Nx

vector of exposure by age

ages

vector of ages that corresponds to the entries of Nx and Dx

num.param

the number of parameters in the model

theta.default

default values for the parameters

theta.range

if not NULL, a range of plausible values that the parameters can assume

optim.default

a list of default parameters to pass into R's optim function

predict.fn

a function which, given values for the parameters and a dataset of exposures, produces predicted numbers of deaths by age. it takes as its arguments:

theta

a vector of parameter values

Nx

vector of exposure by age

age

vector of ages that the Nx correspond to

age.offset

the offset used for age. (See mortalityData-class.)

theta.start.fn

if not NULL, a function which takes a mortalityData object and returns very preliminary estimates of the parameters which can then be used in an optimization routine

simulate.fn

if not NULL, a function which takes as its arguments:

theta

a vector of parameter values

Nx

vector of exposure by age

age

vector of ages that the Nx correspond to

seed

if not NULL, the value to seed the pseudorandom number generator with

means

TODO

eval.fn

if not NULL, a function which produces an evaluation of the model's fit to a given dataset. it takes as its arguments:

fit.obj

the mortalityFit object that resulted from fitting the model to a dataset

observed

the mortalityData object with the observed data

fittedValues

the mortalityPrediction object with fitted values, usually resulting from a call to predict.fn

hazard

if not NULL, the mortalityHazard object used in making this mortalityModel